Teraform/cloud-functions/GetReviews/requirements.txt
functions-framework==3.*
functions-framework==3.*
google-cloud-firestore
Flask